Add some tests of relation data being updated.
Figure out that we have to use a different class for events and charm
because registering with a framework creates attributes on the class
type, and that needs to be in sync with the framework being created.

Also it turns out that Model caches relation data, so even though we
update the value used by the Backend, it doesn't invalidate the
in-memory Relation.data cache. The ugly hack of touching _lazy_data is
the only way forward for now.